Added click upgrade and cleaned up tick
Click upgrade is a new upgrade type that makes clicking gain 1% of cps

Tick function has been simplified with for each loops and the logic works better now

More upgrades have been added for buildings and code has been simplified by having preset values for million, billion, and trillion

Added floating text to show how many cookies are made each click (currently not working)

+++ src/client/java/com/cookie/Upgrade.java
@@ -11,9 +11,9 @@ public class Upgrade {
     public boolean isUnlocked = false;
     public double unlockBuildingCount;
 
-    public Upgrade(String name, double price, String building, double multiplier, double unlockBuildingCount) {
+    public Upgrade(String name, BigDecimal price, String building, double multiplier, double unlockBuildingCount) {
         this.name = name;
-        this.price = BigDecimal.valueOf(price);
+        this.price = price;
         this.building = building;
         this.multiplier = multiplier;
         this.unlockBuildingCount = unlockBuildingCount;
